code review : 이미지 데이터 분리
